fix comments 9285 v2
