+++ app/src/main/java/com/trunghoang/orderhub/data/AuthenticationDataSource.kt
@@ -2,6 +2,6 @@ package com.trunghoang.orderhub.data
 
 import io.reactivex.Single
 
-interface DataSource {
+interface AuthenticationDataSource {
     fun authenticate(email: String, password: String): Single<String>
 }
