+++ app/src/main/java/com/trunghoang/orderhub/data/AuthenticationRepository.kt
@@ -4,6 +4,6 @@ import com.trunghoang.orderhub.data.remote.AuthenticationRemoteDataSource
 import javax.inject.Inject
 
 class AuthenticationRepository @Inject constructor(private val authenticationRemote: AuthenticationRemoteDataSource) {
-    fun callLogin(email: String, password: String) =
+    fun authenticate(email: String, password: String) =
         authenticationRemote.authenticate(email, password)
 }
