+++ app/src/main/java/com/trunghoang/orderhub/di/AuthAPIModule.kt
@@ -1,0 +1,13 @@
+package com.trunghoang.orderhub.di
+
+import com.trunghoang.orderhub.service.AuthHttpClient
+import com.trunghoang.orderhub.service.ServiceGenerator
+import dagger.Module
+import dagger.Provides
+
+@Module
+class AuthAPIModule {
+    @Provides
+    fun provideGHNApi() =
+        ServiceGenerator().createService(AuthHttpClient())
+}
