+++ app/src/main/java/com/trunghoang/orderhub/di/AuthRemoteComponent.kt
@@ -1,0 +1,9 @@
+package com.trunghoang.orderhub.di
+
+import com.trunghoang.orderhub.data.remote.AuthenticationRemoteDataSource
+import dagger.Component
+
+@Component(modules = [AuthAPIModule::class])
+interface AuthRemoteComponent {
+    fun inject(authenticationRemoteDataSource: AuthenticationRemoteDataSource)
+}
