+++ app/src/main/java/com/trunghoang/orderhub/ui/login/LoginViewModel.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import com.trunghoang.orderhub.data.AuthenticationRepository
 import com.trunghoang.orderhub.model.APIResponse
-import com.trunghoang.orderhub.service.LoginService.Companion.KEY_ATTR
-import com.trunghoang.orderhub.service.LoginService.Companion.SELECTOR_TOKEN
+import com.trunghoang.orderhub.service.ServiceGenerator.Companion.KEY_ATTR
+import com.trunghoang.orderhub.service.ServiceGenerator.Companion.SELECTOR_TOKEN
 import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.schedulers.Schedulers
@@ -22,9 +22,9 @@ class LoginViewModel @Inject constructor(
         compositeDisposable.clear()
     }
 
-    fun callLogin(email: String, password: String) {
+    fun authenticate(email: String, password: String) {
         compositeDisposable.add(
-            authenticationRepository.callLogin(email, password)
+            authenticationRepository.authenticate(email, password)
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
                 .doOnSubscribe {
